Fixed compilation errors
